+++ src/commands/backups/restore.py
@@ -17,6 +17,7 @@ from src.utils.cli import ALL_PROFILES_OPTION
 from src.utils.cli import PROFILES_OPTION
 from src.utils.cli import PROJECTS_ARGUMENT
 from src.utils.cli import RUNNING_OPTION
+from src.utils.cli import SERVICES_OPTION
 from src.utils.common import PrintCmdData
 from src.utils.compose_rich import ComposeProject
 from src.utils.compose_rich import get_compose_projects
@@ -74,7 +75,7 @@ def do_restore(
             project.dir,
             project.file,
             project.selected_profiles,
-            command=["down"],
+            command=["down", *project.selected_services],
             dry_run=options.dry_run,
             cmds=cmds,
         )
@@ -89,7 +90,7 @@ def do_restore(
             project.dir,
             project.file,
             project.selected_profiles,
-            command=["up", "-d"],
+            command=["up", "-d", *project.selected_services],
             dry_run=options.dry_run,
             cmds=cmds,
         )
@@ -210,6 +211,7 @@ def get_project_name(project_name: t.Optional[str], project: ComposeProject) -> 
 
 def main(  # noqa: CFQ002 (max arguments)
     projects: list[pathlib.Path] = PROJECTS_ARGUMENT,
+    services: list[str] = SERVICES_OPTION,
     profiles: list[str] = PROFILES_OPTION,
     all_profiles: bool = ALL_PROFILES_OPTION,
     running: bool = RUNNING_OPTION,
@@ -236,6 +238,7 @@ def main(  # noqa: CFQ002 (max arguments)
     compose_projects = list(
         get_compose_projects(
             projects,
+            services,
             all_profiles or profiles,
             ProjectSearchOptions(
                 print_compose_errors=dry_run,