+++ src/commands/log.py
@@ -7,6 +7,7 @@ from src.utils.cli import ALL_PROFILES_OPTION
 from src.utils.cli import PROFILES_OPTION
 from src.utils.cli import PROJECTS_ARGUMENT
 from src.utils.cli import RUNNING_OPTION
+from src.utils.cli import SERVICES_OPTION
 from src.utils.compose_rich import ComposeProject
 from src.utils.compose_rich import get_compose_projects
 from src.utils.compose_rich import ProjectSearchOptions
@@ -30,6 +31,7 @@ def log_project(project: ComposeProject, options: Options, info: ProjectInfo):
             "logs",
             *(["-t"] if options.show_timestamps else []),
             *(["-f"] if options.follow else []),
+            *project.selected_services,
         ],
         dry_run=False,
         cmds=info.cmds,
@@ -37,8 +39,9 @@ def log_project(project: ComposeProject, options: Options, info: ProjectInfo):
     )
 
 
-def main(
+def main(  # noqa: CFQ002 (max arguments)
     projects: list[pathlib.Path] = PROJECTS_ARGUMENT,
+    services: list[str] = SERVICES_OPTION,
     profiles: list[str] = PROFILES_OPTION,
     all_profiles: bool = ALL_PROFILES_OPTION,
     running: bool = RUNNING_OPTION,
@@ -51,6 +54,7 @@ def main(
 
     for project in get_compose_projects(
         projects,
+        services,
         all_profiles or profiles,
         ProjectSearchOptions(
             print_compose_errors=False,